refactor: Update profile and settings layouts
