+++ src/routes/(main)/(protected)/profile/settings/(components)/sidebar-nav.svelte
@@ -16,7 +16,7 @@
   });
 </script>
 
-<nav class={cn("flex space-x-2 lg:flex-col lg:space-x-0 lg:space-y-1", className)}>
+<nav class={cn("flex w-full flex-wrap items-center justify-center gap-2", className)}>
   {#each items as item}
     {@const isActive = $page.url.pathname === item.href}
